Fix the issue of reduction global mem version
In previous implementation, the input data is updated for reduction in-place.
However, this input data is used multiple times, such as warm-up run and getting a average result. Therefore, after the very first run, the input data is not the same any more and the result will be incorrect.

In this fix, the size of output array in the GPU global memory is enlarged to the same size of input data. All the temporary results will be stored in the output array so that the input array keeps read-only.